requirements.txt
faster-whisper >= 1.0.3
sounddevice >= 0.4.7
eel >= 0.17.0
soundfile >= 0.12.1
websockets <= 13.1
librosa >= 0.10.2.post1
openai >= 1.37.0
ollama
pydub
pyaudio
streamlit
python-dotenv
youtube_dl
azure-cognitiveservices-speech